Fix: Event signatures moved to a single file to improve maintainability
